# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Speech-AI Forge API",
    "description": "\n🍦 Speech-AI-Forge 是一个围绕 TTS 生成模型 ChatTTS 开发的项目，实现了 API Server 和 基于 Gradio 的 WebUI。<br/>\n🍦 Speech-AI-Forge is a project developed around the TTS generation model ChatTTS, implementing an API Server and a Gradio-based WebUI.\n\n项目地址: [https://github.com/lenML/Speech-AI-Forge](https://github.com/lenML/Speech-AI-Forge)\n\n> 所有生成音频的 POST api都无法在此页面调试，调试建议使用 playground <br/>\n> All audio generation POST APIs cannot be debugged on this page, it is recommended to use playground for debugging\n\n> 如果你不熟悉本系统，建议从这个一键脚本开始，在colab中尝试一下：<br/>\n> [https://colab.research.google.com/github/lenML/Speech-AI-Forge/blob/main/colab.ipynb](https://colab.research.google.com/github/lenML/Speech-AI-Forge/blob/main/colab.ipynb)\n\n[[Learn More About Documents]](https://github.com/lenML/Speech-AI-Forge/issues/240)\n            ",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/speakers/list": {
      "get": {
        "tags": [
          "Speaker"
        ],
        "summary": "List Speakers",
        "description": "List all available speakers with optional pagination and detail control.\n\n- `detailed`: If true, returns complete metadata including references and embeddings.\n- `offset` / `limit`: Support for paginated speaker listing.",
        "operationId": "list_speakers_v1_speakers_list_get",
        "parameters": [
          {
            "name": "detailed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Detailed"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/speakers/update": {
      "post": {
        "tags": [
          "Speaker"
        ],
        "summary": "Update Speakers",
        "description": "Batch update multiple speakers by providing a list of speaker JSON configs.\n\nEach speaker must already exist (matched by ID).  \nWill overwrite corresponding fields and persist changes to disk.",
        "operationId": "update_speakers_v1_speakers_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeakersUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/speaker/create": {
      "post": {
        "tags": [
          "Speaker"
        ],
        "summary": "Create Speaker",
        "description": "Create a new speaker profile with optional reference audios.\n\n- `name` is required and used as unique identifier.\n- `wavs` is a list of audio samples (base64-encoded) and reference texts for embedding.\n- `save_file`: If true, the speaker will be saved to disk and available after refresh.",
        "operationId": "create_speaker_v1_speaker_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpeaker"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/speaker/update": {
      "post": {
        "tags": [
          "Speaker"
        ],
        "summary": "Update Speaker",
        "description": "Update a single speaker's configuration by full JSON override.\n\nThe speaker must already exist (matched by ID).  \nFields like name, gender, refs, etc., will be updated accordingly.",
        "operationId": "update_speaker_v1_speaker_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpeaker"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/speaker/detail": {
      "post": {
        "tags": [
          "Speaker"
        ],
        "summary": "Speaker Detail",
        "description": "Fetch metadata of a specific speaker by ID.\n\n- `with_emb`: If true, includes embedding vectors and all reference data.",
        "operationId": "speaker_detail_v1_speaker_detail_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeakerDetail"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/prompt/refine": {
      "post": {
        "tags": [
          "Text"
        ],
        "summary": "Refiner Prompt Post",
        "description": "**DeprecationWarning**\n\nThis endpoint is deprecated and will be removed in the future.\n\nRequirements:\n- `chattts` model",
        "operationId": "refiner_prompt_post_v1_prompt_refine_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefineTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/text/normalize": {
      "post": {
        "tags": [
          "Text"
        ],
        "summary": "Text Normalize Post",
        "description": "Normalize raw input text using a selected Text Normalization (TN) pipeline.\n\nThis endpoint supports different TN implementations to perform text normalization \n(e.g., expanding numbers, abbreviations, adding pauses or prosodic markers for TTS, etc.).\n\n### Parameters\n\n- `text` (str): The raw text to normalize.\n- `pipe_id` (str): The TN pipeline to use. Available options:\n  - `base`\n  - `chat-tts`\n  - `cosy-voice`\n  - `fish-speech`\n  - `f5-tts`\n  - `index-tts`\n  - `spark-tts`\n- `config` (TNConfig, optional): Optional configuration to customize TN behavior for specific pipelines.\n\n### Returns\n\nA normalized version of the input text, suitable for use in speech synthesis or downstream NLP tasks.",
        "operationId": "text_normalize_post_v1_text_normalize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextNormalizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioReference": {
        "properties": {
          "wav_b64": {
            "type": "string",
            "title": "Wav B64"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "emotion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emotion",
            "default": "default"
          }
        },
        "type": "object",
        "required": [
          "wav_b64",
          "text"
        ],
        "title": "AudioReference"
      },
...
      "BaseResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "message",
          "data"
        ],
        "title": "BaseResponse"
      },
...
      "CreateSpeaker": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "gender": {
            "type": "string",
            "title": "Gender",
            "default": ""
          },
          "author": {
            "type": "string",
            "title": "Author",
            "default": ""
          },
          "desc": {
            "type": "string",
            "title": "Desc",
            "default": ""
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": ""
          },
          "wavs": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AudioReference"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wavs"
          },
          "save_file": {
            "type": "boolean",
            "title": "Save File",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateSpeaker"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RefineTextRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt",
            "default": "[oral_2][laugh_0][break_6]"
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": -1
          },
          "top_P": {
            "type": "number",
            "title": "Top P",
            "default": 0.7
          },
          "top_K": {
            "type": "integer",
            "title": "Top K",
            "default": 20
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "default": 0.7
          },
          "repetition_penalty": {
            "type": "number",
            "title": "Repetition Penalty",
            "default": 1.0
          },
          "max_new_token": {
            "type": "integer",
            "title": "Max New Token",
            "default": 384
          },
          "spliter_threshold": {
            "type": "integer",
            "title": "Spliter Threshold",
            "default": 300
          },
          "normalize": {
            "type": "boolean",
            "title": "Normalize",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "RefineTextRequest"
      },
...
      "SpeakerDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "with_emb": {
            "type": "boolean",
            "title": "With Emb",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "SpeakerDetail"
      },
...
      "SpeakersUpdate": {
        "properties": {
          "speakers": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Speakers"
          }
        },
        "type": "object",
        "required": [
          "speakers"
        ],
        "title": "SpeakersUpdate"
      },
...
      "TNConfig": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "disabled": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disabled"
          }
        },
        "type": "object",
        "title": "TNConfig"
      },
...
      "TextNormalizeRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "pipe_id": {
            "type": "string",
            "enum": [
              "base",
              "chat-tts",
              "cosy-voice",
              "fish-speech",
              "f5-tts",
              "index-tts",
              "spark-tts"
            ],
            "title": "Pipe Id",
            "default": "base"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TNConfig"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TextNormalizeRequest"
      },
...
      "UpdateSpeaker": {
        "properties": {
          "json": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Json"
          }
        },
        "type": "object",
        "title": "UpdateSpeaker"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:7870"
    },
    {
      "url": "http://0.0.0.0:7870"
    },
    {
      "url": "http://localhost:7870"
    }
  ]
}